Add hint selectors for included countries
