+++ assets/src/css/codefund-framework/_custom.scss
@@ -67,3 +67,11 @@ ul.list-group-kv li.list-group-item {
     height: 10px;
   }
 }
+
+.click-option {
+  cursor: pointer !important;
+}
+
+.click-option:hover {
+  text-decoration: underline;
+}
